+++ worker/src/worker.ts
@@ -3,9 +3,6 @@ import { cors } from 'hono/cors';
 import { jwt } from 'hono/jwt'
 import { Jwt } from 'hono/utils/jwt'
 
-// @ts-ignore
-import { api as apiV1 } from './deprecated';
-
 import { api as commonApi } from './commom_api';
 import { api as mailsApi } from './mails_api'
 import { api as userApi } from './user_api';
@@ -121,12 +118,17 @@ app.route('/', commonApi)
 app.route('/', mailsApi)
 app.route('/', userApi)
 app.route('/', adminApi)
-app.route('/', apiV1)
 app.route('/', apiSendMail)
 app.route('/', telegramApi)
 
-app.get('/', async c => c.text("OK"))
-app.get('/health_check', async c => c.text("OK"))
+app.get('/', async c => {
+	if (!c.env.DB) { return c.text("DB is not available", 400); }
+	return c.text("OK");
+})
+app.get('/health_check', async c => {
+	if (!c.env.DB) { return c.text("DB is not available", 400); }
+	return c.text("OK");
+})
 app.all('/*', async c => c.text("Not Found", 404))